+++ src/main/java/com/project3/placestation/repository/interfaces/ProdReviewRepository.java
@@ -16,7 +16,7 @@ public interface ProdReviewRepository {
 	
 	public int addReview(ProdReview review); // 리뷰 등록
 	public int saveReview(ProdReview review); // 답글 등록
-	public int deleteReview(int prodRevNo);
-	public int countReview(int prodNo);
-	public int avgStar(int prodNo);
+	public int deleteReview(Integer prodRevNo);
+	public int countReview(Integer prodNo);
+	public Double avgStar(Integer prodNo);
 }
